+++ commands/moderation/clearchat.ts
@@ -50,41 +50,51 @@ export default {
                 return;
             }
         }
+        if (!message.guild!.me!.permissions.has([Permissions.FLAGS.MANAGE_MESSAGES])) {
+            clearEmbed
+                .setColor(config.embeds.errorColor as ColorResolvable)
+                .setAuthor({ name: 'No tengo permisos para realizar esta acción.', iconURL: 'attachment://error-icon.png' })
+            message.reply({ embeds: [clearEmbed], files: [errorImg] });
+            return;
+        }
 
         await message.delete();
         let fetched;
         let filtered;
         let limit;
-        let counter = 0;
         do {
             totalLimit > 100 ? limit = 100 : limit = totalLimit;
             fetched = await message.channel.messages.fetch({ limit: limit });
+            console.log(`fetch ${fetched.size}`);
             if (member !== undefined) {
                 filtered = fetched.filter(msg => Date.now() - msg.createdTimestamp < 1209600000 && msg.member! === member);
             } else {
                 filtered = fetched.filter(msg => Date.now() - msg.createdTimestamp < 1209600000);
             }
-            counter += filtered.size;
-            await message.channel.bulkDelete(filtered);
-            if (totalLimit >= filtered.size) break;
+            try {
+                await message.channel.bulkDelete(filtered, true);
+            } catch (error) {
+                continue;
+            }
+            if (filtered.size < 100) break;
             totalLimit -= limit;
         } while (totalLimit > 0)
 
         let msg = 'mensaje';
-        if (filtered.size > 1 || filtered.size === 0) {
+        if (filtered.size > 1 || filtered.size < 1) {
             msg = 'mensajes';
         }
         if (fetched.some((msg: { createdTimestamp: number; }) => Date.now() - msg.createdTimestamp > 1209600000)) {
             clearEmbed
                 .setColor(config.embeds.successColor as ColorResolvable)
-                .setAuthor({ name: `He borrado ${counter} ${msg}. \n`, iconURL: 'attachment://success-icon.png' })
+                .setAuthor({ name: `Mensajes eliminados`, iconURL: 'attachment://success-icon.png' })
                 .setDescription('Debido a las limitaciones de Discord, no puedo eliminar mensajes que tengan más de 14 días.')
             message.channel.send({ embeds: [clearEmbed], files: [successImg] }).then(msg => setTimeout(() => msg.delete(), 5000)).catch();
             return;
         }
         clearEmbed
             .setColor(config.embeds.successColor as ColorResolvable)
-            .setAuthor({ name: `He borrado ${counter} ${msg}. \n`, iconURL: 'attachment://success-icon.png' })
+            .setAuthor({ name: `Mensajes eliminados`, iconURL: 'attachment://success-icon.png' })
         message.channel.send({ embeds: [clearEmbed], files: [successImg] }).then(msg => setTimeout(() => msg.delete(), 5000)).catch();
     }
 }
